fix types

+++ Source/Raspicam.Net/CurrentState.cs
@@ -14,7 +14,7 @@ namespace Raspicam.Net
 
         public double AnalogGain { get => _camera.GetAnalogGain(); set => _camera.SetAnalogGain(value); }
         public double DigitalGain { get => _camera.GetDigitalGain(); set => _camera.SetDigitalGain(value); }
-        public int ShuttderSpeedMs { get => _camera.GetShutterSpeed(); set => _camera.SetShutterSpeed(value); }
-        public uint Iso { get => _camera.GetIso(); set => _camera.SetIso(value); }
+        public uint ShuttderSpeedMs { get => _camera.GetShutterSpeed(); set => _camera.SetShutterSpeed((int)value); }
+        public uint Iso { get => _camera.GetIso(); set => _camera.SetIso((int)value); }
     }
 }
